Refactored name sanitation
