+++ src/MinIOS3Provider/MinIOS3Provider.Test/NamingTest.cs
@@ -8,17 +8,11 @@ namespace MinIOS3Provider.Test
     [TestFixture]
     public class Naming_Tests
     {
-        private BucketController _controller;
-        private Mock<IMinioS3Client> _client;
-        private Mock<ILogger<BucketController>> _logger;
 
         [SetUp]
         public void Setup()
         {
-            _client = new Mock<IMinioS3Client>();
-            _logger = new Mock<ILogger<BucketController>>();
 
-            _controller = new BucketController(_logger.Object, _client.Object);
         }
 
         [Test]
@@ -26,8 +20,7 @@ namespace MinIOS3Provider.Test
         [TestCase("this-is-a-valid-name-and-this")]
         public void Bucket_names_unchanged(string name)
         {
-            //var name = "this-is-a-valid-name";
-            var newName = _controller.SanitizeBucketName(name);
+            var newName = NameUtil.SanitizeBucketName(name);
             Assert.AreEqual(name, newName);
         }
 
@@ -39,8 +32,7 @@ namespace MinIOS3Provider.Test
         [TestCase("xn--test")]
         public void Bucket_names_changed(string name)
         {
-            //var name = "this-is-a-valid-name";
-            var newName = _controller.SanitizeBucketName(name);
+            var newName = NameUtil.SanitizeBucketName(name);
             Assert.AreNotEqual(name, newName);
         }
 
@@ -53,9 +45,8 @@ namespace MinIOS3Provider.Test
         [TestCase("xn--test")]
         public void Bucket_names_ShouldPassSecondTime(string name)
         {
-            //var name = "this-is-a-valid-name";
-            name = _controller.SanitizeBucketName(name);
-            var newName = _controller.SanitizeBucketName(name);
+            name = NameUtil.SanitizeBucketName(name);
+            var newName = NameUtil.SanitizeBucketName(name);
             Assert.AreEqual(name, newName);
         }
     }
